# kpc-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/topic/{username}/list": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Topics",
        "operationId": "list_topics_topic__username__list_get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Topics Topic  Username  List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topic/keyphrase_clustering/{username}/{topic}/{keyphrase_order}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Keyphrase Clusters",
        "operationId": "list_keyphrase_clusters_topic_keyphrase_clustering__username___topic___keyphrase_order__get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          },
          {
            "name": "keyphrase_order",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Keyphrase Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Keyphrase Clusters Topic Keyphrase Clustering  Username   Topic   Keyphrase Order  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topic/keyphrase_clustering/{username}/{topic}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Keyphrase Clusters",
        "operationId": "list_keyphrase_clusters_topic_keyphrase_clustering__username___topic__get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          },
          {
            "name": "keyphrase_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Keyphrase Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Keyphrase Clusters Topic Keyphrase Clustering  Username   Topic  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topic/clusters/{username}/{topic}/{cluster_order}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Clusters",
        "operationId": "list_clusters_topic_clusters__username___topic___cluster_order__get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          },
          {
            "name": "cluster_order",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Clusters Topic Clusters  Username   Topic   Cluster Order  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topic/clusters/{username}/{topic}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Clusters",
        "operationId": "list_clusters_topic_clusters__username___topic__get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          },
          {
            "name": "cluster_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Cluster Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Clusters Topic Clusters  Username   Topic  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/topic/cluster_selection/{username}/{topic}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "List Cluster Selection",
        "operationId": "list_cluster_selection_topic_cluster_selection__username___topic__get",
        "security": [
          {
            "OAuth2PasswordBearerCookie": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Cluster Selection Topic Cluster Selection  Username   Topic  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearerCookie": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/users/token"
          }
        }
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}